worker/summarizer/requirements.txt
openai
python-dotenv
kafka-python